chore: clean directory structure
